Module aws_sdk_synthetics::model

Modules

base_screenshot

See BaseScreenshot

canary

See Canary

canary_code_input

See CanaryCodeInput

canary_code_output

See CanaryCodeOutput

canary_last_run

See CanaryLastRun

canary_run

See CanaryRun

canary_run_config_input

See CanaryRunConfigInput

canary_run_config_output

See CanaryRunConfigOutput

canary_run_status

See CanaryRunStatus

canary_run_timeline

See CanaryRunTimeline

canary_schedule_input

See CanaryScheduleInput

canary_schedule_output

See CanaryScheduleOutput

canary_status

See CanaryStatus

canary_timeline

See CanaryTimeline

runtime_version

See RuntimeVersion

visual_reference_input

See VisualReferenceInput

visual_reference_output

See VisualReferenceOutput

vpc_config_input

See VpcConfigInput

vpc_config_output

See VpcConfigOutput

Structs

BaseScreenshot

A structure representing a screenshot that is used as a baseline during visual monitoring comparisons made by the canary.

Canary

This structure contains all information about one canary in your account.

CanaryCodeInput

Use this structure to input your script code for the canary. This structure contains the

CanaryCodeOutput

This structure contains information about the canary's Lambda handler and

CanaryLastRun

This structure contains information about the most recent run of a single canary.

CanaryRun

This structure contains the details about one run of one canary.

CanaryRunConfigInput

A structure that contains input information for a canary run.

CanaryRunConfigOutput

A structure that contains information about a canary run.

CanaryRunStatus

This structure contains the status information about a canary run.

CanaryRunTimeline

This structure contains the start and end times of a single canary run.

CanaryScheduleInput

This structure specifies how often a canary is to make runs and the date and time

CanaryScheduleOutput

How long, in seconds, for the canary to continue making regular runs according to the schedule in the

CanaryStatus

A structure that contains the current state of the canary.

CanaryTimeline

This structure contains information about when the canary was created and modified.

RuntimeVersion

This structure contains information about one canary runtime version. For more information about

VisualReferenceInput

An object that specifies what screenshots to use as a baseline for visual monitoring by this canary, and optionally the parts of the screenshots to ignore during the visual monitoring comparison.

VisualReferenceOutput

If this canary performs visual monitoring by comparing screenshots, this structure contains the ID of the canary run that is used as the baseline for screenshots, and the coordinates

VpcConfigInput

If this canary is to test an endpoint in a VPC, this structure contains

VpcConfigOutput

If this canary is to test an endpoint in a VPC, this structure contains

Enums

CanaryRunState

CanaryRunStateReasonCode

CanaryState

CanaryStateReasonCode